+++ src/ct_physics.jl
@@ -607,6 +607,43 @@ function breaction!(f, u, bnode, data, ::Type{InterfaceRecombination})
     return
 end
 
+###########################################################################
+###########################################################################
+
+
+"""
+$(TYPEDSIGNATURES)
+Creates boundary conditions for gate contacts. A Robin boundary condition is applied to the electrostatic potential
+    
+``\\varepsilon_\\mathrm{s} \\nabla \\psi \\cdot \\nu + \\frac{\\varepsilon_\\mathrm{ox}}{d_\\mathrm{ox}} (\\psi - U_G) = Q_{ss}``,
+
+where ``\\varepsilon_\\mathrm{ox}`` denotes the absolute dielectric permittivity of the oxide and ``d_\\mathrm{ox}`` the thickness of the oxide.
+The term ``Q_{ss}`` corresponds to the surface charge density at the gate contact.
+
+For the quasi Fermi potentials, homogeneous Neumann boundary conditions are implemented.
+
+Note that an additional reference voltage ``U_{ref}`` can be absorbed into the surface charge term.
+This leads to an effective surface charge density
+
+``Q_{ss}^{'} = Q_{ss} + \\frac{\\varepsilon_\\mathrm{ox}}{d_\\mathrm{ox}} U_{ref}``.
+
+Equivalently, in terms of surface state density,
+
+``Q_{ss}^{'} = q N_{ss}^{'} ``.
+"""
+
+function breaction!(f, u, bnode, data, ::Type{GateContact})
+
+    params = data.params
+    ipsi = data.index_psi
+
+    # Homogeneous Neumann boundary conditions for electrons and holes by default
+    # Robin boundary condition for the electrostatic potential
+    f[ipsi] = (params.dielectricConstantOxideGate[bnode.region] / params.thicknessOxideGate[bnode.region]) * (u[ipsi] - params.contactVoltage[bnode.region]) - params.surfaceChargeDensityGate[bnode.region]
+
+    return
+end
+
 ##########################################################
 ##########################################################
 
@@ -1630,9 +1667,18 @@ function addTrapCaptureEscape!(f, u, node, data, ::Type{NoTrap})
 end
 """
 $(TYPEDSIGNATURES)
-A simple trap with one state that can either be filled or empty
+Recombination with a trap with one state that can either be filled or empty
+
+The reaction rate is give by:\\
+R ∝ sₙ (1 - f) - eₙ f [if the trap and band have the same charge] \\
+R ∝ sₙ f - eₙ (1 - f) [if the trap and band have opposite charges].
+
+sₙ and eₙ are the capture and escape rate, which are related via detailed balance:\\
+eₙ = sₙ Nc γ exp( zc (Ec - Et) / kT ).
+
+The formalism works when the trap statistics are treated with FermiDiracMinusOne, GaussFermiPaasch or GaussFermiSimpson13.
 """
-function addTrapCaptureEscape!(f, u, node, data, ::Type{SingleStateTrap})
+function addTrapCaptureEscape!(f, u, node, data, ::Type{TrapCaptureEscape})
 
     (; k_B, q) = data.constants
 
@@ -1646,11 +1692,8 @@ function addTrapCaptureEscape!(f, u, node, data, ::Type{SingleStateTrap})
         # Account for non-Boltzmann statistics in detailed balance to compute escape rate.
         # It is assumed that the trap is described using FermiDiracMinusOne. The correction
         # is of F(η)/exp(η) where F is the function used in the carrier state equation.
-        if (data.F[icc] == Boltzmann)
-            nonBoltzmannReductionFactor = 1.0
-        else
-            nonBoltzmannReductionFactor = ncc / (Nc * exp(etaFunction!(u, node, data, icc)))
-        end
+        nonBoltzmannReductionFactor = ncc / (Nc * exp(etaFunction!(u, node, data, icc)))
+
 
         Ec = data.params.bandEdgeEnergy[icc]
         zc = data.params.chargeNumbers[icc]
@@ -1662,15 +1705,12 @@ function addTrapCaptureEscape!(f, u, node, data, ::Type{SingleStateTrap})
                 itc = data.chargeCarrierList[itc] # find correct index within chargeCarrierList
                 s = capture[itc, icc, node.region]
 
-                if (s > 0) # Only compute where there is capture
+                if s > 0 # Only compute where there is capture
                     zt = data.params.chargeNumbers[itc]
 
                     ntc = get_density!(u, node, data, itc)
                     Nt = data.params.densityOfStates[itc]
                     Et = data.params.bandEdgeEnergy[itc]
-
-                    # Escape computed from detailed balance assuming traps described using FD-minus one
-                    e = s * Nc * exp(zc * (Ec - Et) / (k_B * T)) * nonBoltzmannReductionFactor
 
                     # Allow for both acceptor and donor trap in one line
                     # e.g.  If acceptor traps (trap charge = -1) then reaction with
@@ -1678,9 +1718,13 @@ function addTrapCaptureEscape!(f, u, node, data, ::Type{SingleStateTrap})
                     # the valence band is r = Nt*( s*p*f - e*(1-f) ).
                     # For donor traps the (1-f) and f swaps, which is done
                     # using sign(zc*zt).
-                    captureFactor = (sign(zc * zt) + 1) / 2 - sign(zc * zt) * ntc / Nt
-                    escapeFactor = 1 - captureFactor
-                    r = Nt * (s * ncc * captureFactor - e * escapeFactor)
+                    occupationFactor = (sign(zc * zt) + 1) / 2 + sign(-zc * zt) * ntc / Nt
+
+                    # The reaction rate is rewritten in a more convenient form, namely (zc=zt)
+                    # r = Nt s * (1-f(η)) * ( 1 - exp( zc * q/kBT * (φₜ - φₙ) ) )
+                    # or (zc=-zt)
+                    # r = Nt s * f(η) * ( 1 - exp( zc * q/kBT * (φₜ - φₙ) ) )
+                    r = Nt * (s * ncc * occupationFactor) * (1.0 - exp(zc / (k_B * T) * q * (u[itc] - u[icc])))
 
                     # For the reaction expression we use the charge of the band as (e.g.) holes can enter
                     # an electron trap from the valence band, and using the trap charge would not capture this.
@@ -1692,173 +1736,6 @@ function addTrapCaptureEscape!(f, u, node, data, ::Type{SingleStateTrap})
     end
 
     return
-end
-"""
-$(TYPEDSIGNATURES)
-A Gaussian distribution of traps with one state that can either be filled or empty
-"""
-function addTrapCaptureEscape!(f, u, node, data, ::Type{GaussianDistributedTrap})
-
-    (; k_B, q) = data.constants
-
-    capture = data.params.recombinationTrapCaptureRates
-    σ_T = data.params.trapDistributionWidth
-    T = data.params.temperature
-    
-
-    for icc in data.chargeCarrierList
-        ncc = get_density!(u, node, data, icc)
-        Nc = data.params.densityOfStates[icc]
-
-        # Account for non-Boltzmann statistics in detailed balance to compute escape rate.
-        # It is assumed that the trap is described using FermiDiracMinusOne. The correction
-        # is of F(η)/exp(η) where F is the function used in the carrier state equation.
-        if (data.F[icc] == Boltzmann)
-            nonBoltzmannReductionFactor = 1.0
-        else
-            nonBoltzmannReductionFactor = ncc / (Nc * exp(etaFunction!(u, node, data, icc)))
-        end
-
-        Ec = data.params.bandEdgeEnergy[icc]
-        zc = data.params.chargeNumbers[icc]
-
-        for iitc in data.trapCarrierList
-            # add trap carriers only in defined regions (otherwise get NaN error)
-            if node.region ∈ iitc.regions
-                itc = iitc.trapCarrier            # species number chosen by user
-                itc = data.chargeCarrierList[itc] # find correct index within chargeCarrierList
-
-                GFI = data.F[itc]
-                s = capture[itc, icc, node.region]
-
-                if (s > 0) # Only compute where there is capture
-
-                    Nt = data.params.densityOfStates[itc]
-                    Et = data.params.bandEdgeEnergy[itc]
-
-                    # Compute capture rate by integrating over Gaussian trap distribution
-                    βc=etaFunction!(u, node, data, itc)
-                    ξ=sqrt(2)*σ_T[itc]/(k_B*T)
-                    captureIntegral = GFI(βc)
-                    # I = distributedTrapsIntegral(exp(-βc), ξ)
-                    # if(100*abs(captureIntegral - I)/I > 2)
-                    #     println(100*abs(captureIntegral - I)/I)
-                    # end
-                    zt = data.params.chargeNumbers[itc]
-
-
-                    # Escape computed from detailed balance assuming traps described using FD-minus one
-                    # e = s * Nc * exp(zc * (Ec - Et) / (k_B * T)) * nonBoltzmannReductionFactor
-                    βe=βc - σ_T[itc]^2/(k_B*T)^2 # Minus sign because of definition of eta and distribution function?
-                    escapeIntegral = GFI(βe)
-                    e = s * Nc * exp(zc * (Ec - Et) / (k_B * T)) * nonBoltzmannReductionFactor
-                    e *= exp( 0.5*(σ_T[itc]/(k_B*T))^2 ) 
-
-                    # I = distributedTrapsIntegral(exp(-βc + σ_T[itc,node.region]^2/(k_B*T)^2), ξ)
-                    # if(100*abs(escapeIntegral - I)/I > 5)
-                    #     if(maximum([I,escapeIntegral])>1e-16)
-                    #         println(100*abs(escapeIntegral - I)/I, " " , I)
-                    #     end
-                    # end
-                    # Allow for both acceptor and donor trap in one line
-                    captureFactor = (sign( zc * zt) + 1) / 2 - sign( zc * zt) * captureIntegral
-                    escapeFactor  = (sign(-zc * zt) + 1) / 2 - sign(-zc * zt) * escapeIntegral
-                    r = Nt * (s * ncc * captureFactor - e * escapeFactor)
-
-                    # For the reaction expression we use the charge of the band as (e.g.) holes can enter
-                    # an electron trap from the valence band, and using the trap charge would not capture this.
-                    f[icc] = f[icc] + q * zc * r    #
-                    f[itc] = f[itc] - q * zc * r    #
-                end
-            end
-        end
-    end
-
-    return
-end
-# function addTrapCaptureEscape!(f, u, node, data, ::Type{GaussianDistributedTrap})
-
-#     (; k_B, q) = data.constants
-
-#     capture = data.params.recombinationTrapCaptureRates
-#     σ_T = data.params.trapDistributionWidth
-#     T = data.params.temperature
-
-#     for icc in data.chargeCarrierList
-#         ncc = get_density!(u, node, data, icc)
-#         Nc = data.params.densityOfStates[icc]
-
-#         # Account for non-Boltzmann statistics in detailed balance to compute escape rate.
-#         # It is assumed that the trap is described using FermiDiracMinusOne. The correction
-#         # is of F(η)/exp(η) where F is the function used in the carrier state equation.
-#         if (data.F[icc] == Boltzmann)
-#             nonBoltzmannReductionFactor = 1.0
-#         else
-#             nonBoltzmannReductionFactor = ncc / (Nc * exp(etaFunction!(u, node, data, icc)))
-#         end
-
-#         Ec = data.params.bandEdgeEnergy[icc]
-#         zc = data.params.chargeNumbers[icc]
-
-#         for iitc in data.trapCarrierList
-#             # add trap carriers only in defined regions (otherwise get NaN error)
-#             if node.region ∈ iitc.regions
-#                 itc = iitc.trapCarrier            # species number chosen by user
-#                 itc = data.chargeCarrierList[itc] # find correct index within chargeCarrierList
-
-#                 s = capture[itc, icc, node.region]
-
-#                 if (s > 0) # Only compute where there is capture
-
-#                     Nt = data.params.densityOfStates[itc]
-#                     Et = data.params.bandEdgeEnergy[itc]
-
-#                     # Compute capture rate by integrating over Gaussian trap distribution
-#                     βc=exp(-etaFunction!(u, node, data, itc))
-#                     ξ=sqrt(2)*σ_T[itc,node.region]/(k_B*T)
-#                     captureIntegral = distributedTrapsIntegral(βc, ξ)
-#                     zt = data.params.chargeNumbers[itc]
-
-
-#                     # Escape computed from detailed balance assuming traps described using FD-minus one
-#                     # e = s * Nc * exp(zc * (Ec - Et) / (k_B * T)) * nonBoltzmannReductionFactor
-#                     βe=βc*exp(σ_T[itc,node.region]^2/(k_B*T)^2)
-#                     escapeIntegral = distributedTrapsIntegral(βe, ξ)
-#                     e = s * Nc * exp(zc * (Ec - Et) / (k_B * T)) * nonBoltzmannReductionFactor
-#                     e *= exp( 0.5*(σ_T[itc,node.region]/(k_B*T))^2 ) 
-
-#                     # Allow for both acceptor and donor trap in one line
-#                     captureFactor = (sign( zc * zt) + 1) / 2 - sign( zc * zt) * captureIntegral
-#                     escapeFactor  = (sign(-zc * zt) + 1) / 2 - sign(-zc * zt) * escapeIntegral
-#                     r = Nt * (s * ncc * captureFactor - e * escapeFactor)
-
-#                     # For the reaction expression we use the charge of the band as (e.g.) holes can enter
-#                     # an electron trap from the valence band, and using the trap charge would not capture this.
-#                     f[icc] = f[icc] + q * zc * r    #
-#                     f[itc] = f[itc] - q * zc * r    #
-#                 end
-#             end
-#         end
-#     end
-
-#     return
-# end
-"""
-$(TYPEDSIGNATURES)
-Function to compute integrals of the form int e^-x^2/(1+βe^ξx)
-which show up in the Gaussian trap model. Upper and lower bounds are fixed to +-6
-"""
-function distributedTrapsIntegral(β, ξ) #::Float64, ξ::Float64)
-    I = 0
-    xmax=6.0
-    xmin=-xmax
-    nPoints=1000
-    dX = (xmax-xmin)/(nPoints-1)
-    for i=1:nPoints
-        x= xmin + dX*(i-1)
-        I += exp(-x^2)/(1+β*exp(ξ*x))
-    end
-    return dX/√π * I
 end
 """
 $(TYPEDSIGNATURES)